Struct tor_config::CfgPath

pub struct CfgPath(/* private fields */);

A path in a configuration file: tilde expansion is performed, along with expansion of certain variables.

The supported variables are:

ARTI_CACHE: an arti-specific cache directory.

ARTI_CONFIG: an arti-specific configuration directory.

ARTI_SHARED_DATA: an arti-specific directory in the user’s “shared data” space.

ARTI_LOCAL_DATA: an arti-specific directory in the user’s “local data” space.

PROGRAM_DIR: the directory of the currently executing binary. See documentation for std::env::current_exe for security notes.

USER_HOME: the user’s home directory.

These variables are implemented using the directories crate, and so should use appropriate system-specific overrides under the hood. (Some of those overrides are based on environment variables.) For more information, see that crate’s documentation.

Alternatively, a CfgPath can contain literal PathBuf, which will not be expanded.

Implementations

impl CfgPath

pub fn new(s: String) -> Self

Create a new configuration path

pub fn new_literal<P: Into<PathBuf>>(path: P) -> Self

Construct a new CfgPath designating a literal not-to-be-expanded PathBuf

pub fn path(&self) -> Result<PathBuf, CfgPathError>

Return the path on disk designated by this CfgPath.

pub fn as_unexpanded_str(&self) -> Option<&str>

If the CfgPath is a string that should be expanded, return the (unexpanded) string,

Before use, this string would have be to expanded. So if you want a path to actually use, call path instead.

Returns None if the CfgPath is a literal PathBuf not intended for expansion.

pub fn as_literal_path(&self) -> Option<&Path>

If the CfgPath designates a literal not-to-be-expanded Path, return a reference to it

Returns None if the CfgPath is a string which should be expanded, which is the usual case.