Trait tor_config::load::TopLevel

pub trait TopLevel {
    type Builder: DeserializeOwned;

    const DEPRECATED_KEYS: &'static [&'static str] = _;
}

Top-level configuration struct, made from a deserializable builder

One configuration consumer’s configuration settings.

Implementing this trait only for top-level configurations, which are to be parsed at the root level of a (TOML) config file taxonomy.

This trait exists to:

Mark the toplevel configuration structures as suitable for use with resolve

Provide the type of the Builder for use by Rust generic code

Required Associated Types

type Builder: DeserializeOwned

The Builder which can be used to make a Self

Should satisfy &'_ Self::Builder: Builder<Built=Self>

Provided Associated Constants

const DEPRECATED_KEYS: &'static [&'static str] = _

Deprecated config keys, as “.”-separates strings

Object Safety

This trait is not object safe.